fix: remove redundant check on startStage for previous step completion.
in the startStage function, it is expected that it will be called
to kick off a stage which in turn completes the previous stage (if
it is not stage 1).

the check being removed would not allow it to get to the check where
it would allow for the first stage to be completed.